+++ src/main.rs
@@ -14,7 +14,10 @@ extern crate log;
 use clap::{Arg, App, SubCommand};
 
 mod say;
+mod format;
 mod cmd;
+
+use format::{OutputFormat};
 
 fn main() {
     env_logger::init().unwrap();
@@ -39,7 +42,7 @@ fn main() {
                                                       .help("Output format")
                                                       .long("format")
                                                       .value_name("format")
-                                                      .possible_values(&["csv"])
+                                                      .possible_values(&["csv", "json"])
                                                       .required(true))
                                              .arg(Arg::with_name("output")
                                                       .help("Write output to <file>")
@@ -63,10 +66,15 @@ fn main() {
         let state = matches.value_of("state").unwrap_or("all").to_owned();
         let repopath = matches.value_of("repopath").unwrap().to_owned();
         let oauth_token = matches.value_of("oauth-token").unwrap().to_owned();
-        let format = matches.value_of("format").unwrap().to_owned();
+        let format = value_t!(matches, "format", OutputFormat).unwrap_or_else(|e| e.exit());
         let output = matches.value_of("output").unwrap().to_owned();
 
-        cmd::fetch::run(repopath, oauth_token, labels, state, format, output);
+        cmd::fetch::run(repopath,
+                        oauth_token,
+                        labels,
+                        state,
+                        format,
+                        output);
     }
 }