+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## next
+
+* Add JSON output format.
+
 ## 0.2.0
 
 * Change `--csv` to `--format=csv`.